refactoring!
