+++ internal/bus/struct.go
@@ -5,6 +5,12 @@ import (
 	. "events"
 )
 
+type Bus struct {
+	Channels   map[int](chan map[string]any)
+	Events     EventList
+	Dispatcher Dispatcher
+}
+
 func NewBus(dispatcher *Dispatcher) *Bus {
 	channel := make(map[int](chan map[string]any))
