--- internal/bus/type.go
@@ -1,12 +1,0 @@
-package bus
-
-import (
-	. "dispatcher"
-	. "events"
-)
-
-type Bus struct {
-	Channels   map[int](chan map[string]any)
-	Events     EventList
-	Dispatcher Dispatcher
-}
