+++ internal/dispatcher/struct.go
@@ -9,7 +9,3 @@ type Data struct {
 	Metadata string
 }
 type Dispatcher map[reflect.Type]func(EventArgs) (Data, error)
-
-func (dispatcher *Dispatcher) Get(projection any) func(EventArgs) (Data, error) {
-	return (*dispatcher)[reflect.TypeOf(projection)]
-}
