+++ internal/events/struct.go
@@ -1,5 +1,15 @@
 package events
 
+type EventArgs []any
+
+type Event struct {
+	Projection any
+	EventArgs  EventArgs
+	Status     bool
+}
+
+type EventList []Event
+
 func NewEvent(projection any, args EventArgs) *Event {
 	return &Event{
 		Projection: projection,
