--- internal/events/types.go
@@ -1,11 +1,0 @@
-package events
-
-type EventArgs []any
-
-type Event struct {
-	Projection any
-	EventArgs  EventArgs
-	Status     bool
-}
-
-type EventList []Event
